refactor(nvim): restructure lua configuration files
