+++ src/settings/.config/nvim/init.lua
@@ -1,6 +1,3 @@
-require('plugin-settings')
-if require('utils').file_exists(vim.fn.stdpath('config') .. '/lua/custom-settings.lua') then
-  require('custom-settings')
-end
-require('settings')
-require('plugins')
+require('config')
+require('config.options')
+require('config.lazy')
