+++ src/settings/.config/nvim/lua/config/options.lua
