+++ src/settings/.config/nvim/lua/plugins/editor.lua
@@ -1,0 +1,52 @@
+return {
+  {
+    'numToStr/Comment.nvim',
+    dependencies = {
+      'nvim-treesitter/nvim-treesitter',
+      "JoosepAlviste/nvim-ts-context-commentstring",
+    },
+    opts = function()
+      local ft = require('Comment.ft')
+      local m = require("config").editor.comment
+
+      ft({ 'verilog', 'verilog_systemverilog' }, ft.get('c'))
+
+      return {
+        ---LHS of toggle mappings in NORMAL mode
+        toggler = {
+          ---Line-comment toggle keymap
+          line = m.toggle_line,
+          ---Block-comment toggle keymap
+          block = m.toggle_block,
+        },
+        ---LHS of operator-pending mappings in NORMAL and VISUAL mode
+        opleader = {
+          ---Line-comment keymap
+          line = m.opleader_line,
+          ---Block-comment keymap
+          block = m.opleader_block,
+        },
+        ---LHS of extra mappings
+        extra = {
+          ---Add comment on the line above
+          above = m.add_above,
+          ---Add comment on the line below
+          below = m.add_below,
+          ---Add comment at the end of line
+          eol = m.add_eol,
+        },
+        pre_hook = require('ts_context_commentstring.integrations.comment_nvim').create_pre_hook(),
+      }
+
+    end,
+    keys = function()
+      local m = require("config").editor.comment
+      local keys = {}
+      for _, key in pairs(m) do
+        table.insert(keys, key)
+      end
+
+      return keys
+    end,
+  },
+}
