+++ src/settings/.config/nvim/lua/plugins/navigation.lua
@@ -1,0 +1,229 @@
+return {
+  {
+    "henry-hsieh/flash.nvim",
+    dependencies = {
+      {
+        "folke/snacks.nvim",
+        opts = {
+          picker = {
+            win = {
+              input = {
+                keys = {
+                  ["<a-s>"] = { "flash", mode = { "n", "i" } },
+                  ["s"] = { "flash" },
+                },
+              },
+            },
+            actions = {
+              flash = function(picker)
+                require("flash").jump({
+                  pattern = "^",
+                  label = { after = { 0, 0 } },
+                  search = {
+                    mode = "search",
+                    exclude = {
+                      function(win)
+                        return vim.bo[vim.api.nvim_win_get_buf(win)].filetype ~= "snacks_picker_list"
+                      end,
+                    },
+                  },
+                  action = function(match)
+                    local idx = picker.list:row2idx(match.pos[1])
+                    picker.list:_move(idx, true, true)
+                  end,
+                })
+              end,
+            },
+          },
+        },
+      },
+    },
+    branch = 'feat-unique-label-win',
+    ---@type Flash.Config
+    opts = {
+      labels = "asdfghjklqwertyuiopzxcvbnm",
+      search = {
+        exclude = {
+          "notify",
+          "cmp_menu",
+          "noice",
+          "flash_prompt",
+          function(win)
+            -- exclude non-focusable windows
+            return not vim.api.nvim_win_get_config(win).focusable
+          end,
+        },
+      },
+      label = {
+        uppercase = false,
+      },
+    },
+    -- stylua: ignore
+    keys = {
+      { "s",     mode = { "n", "x", "o" }, function() require("flash").jump() end,              desc = "Flash" },
+      { "S",     mode = { "n", "x", "o" }, function() require("flash").treesitter() end,        desc = "Flash Treesitter" },
+      { "r",     mode = "o",               function() require("flash").remote() end,            desc = "Remote Flash" },
+      { "R",     mode = { "o", "x" },      function() require("flash").treesitter_search() end, desc = "Treesitter Search" },
+      { "<c-s>", mode = { "c" },           function() require("flash").toggle() end,            desc = "Toggle Flash Search" },
+      {
+        "<leader>w",
+        mode = { "n", "x", "o" },
+        function()
+          require("utils").flash_jump_regex([[\<]])
+        end,
+        desc = "Jump to word start"
+      },
+      {
+        "<leader>W",
+        mode = { "n", "x", "o" },
+        function()
+          require("utils").flash_jump_regex([[\(^\|\s\+\)\zs\S\ze\S*]])
+        end,
+        desc = "Jump to WORD start"
+      },
+      {
+        "<leader>e",
+        mode = { "n", "x", "o" },
+        function()
+          require("utils").flash_jump_regex([[\w\>]])
+        end,
+        desc = "Jump to word end"
+      },
+      {
+        "<leader>E",
+        mode = { "n", "x", "o" },
+        function()
+          require("utils").flash_jump_regex([[\(^\|\s\+\)\S*\zs\S\ze]])
+        end,
+        desc = "Jump to WORD end"
+      },
+      {
+        "<leader>l",
+        mode = { "n", "x", "o" },
+        function()
+          require("utils").flash_jump_regex([[^\s*\zs\(\S\|$\)\ze]])
+        end,
+        desc = "Jump to line start"
+      },
+      {
+        "<leader>L",
+        mode = { "n", "x", "o" },
+        function()
+          require("utils").flash_jump_regex([[.\=$]])
+        end,
+        desc = "Jump to line end"
+      },
+    },
+  },
+
+  {
+    "folke/trouble.nvim",
+    dependencies = {
+      {
+        "nvim-lualine/lualine.nvim",
+        opts = function(_, opts)
+          -- trouble.nvim
+          local trouble = require("trouble")
+          local symbols = trouble.statusline({
+            mode = "lsp_document_symbols",
+            groups = {},
+            title = false,
+            filter = { range = true },
+            format = "{kind_icon}{symbol.name:Normal}",
+            -- The following line is needed to fix the background color
+            -- Set it to the lualine section you want to use
+            hl_group = "lualine_c_normal",
+          })
+          opts.sections = opts.sections or {}
+          opts.sections.lualine_c = opts.sections.lualine_c or {}
+          table.insert(opts.sections.lualine_c, {
+            symbols.get,
+            cond = symbols.has,
+            separator = "",
+          })
+        end,
+      },
+    },
+    opts = {}, -- for default options, refer to the configuration section for custom setup.
+    cmd = "Trouble",
+    keys = {
+      {
+        "<leader>xx",
+        "<cmd>Trouble diagnostics toggle<cr>",
+        desc = "Diagnostics (Trouble)",
+      },
+      {
+        "<leader>xX",
+        "<cmd>Trouble diagnostics toggle filter.buf=0<cr>",
+        desc = "Buffer Diagnostics (Trouble)",
+      },
+      {
+        "<leader>cs",
+        "<cmd>Trouble symbols toggle focus=false<cr>",
+        desc = "Symbols (Trouble)",
+      },
+      {
+        "<leader>cl",
+        "<cmd>Trouble lsp toggle focus=false win.position=right<cr>",
+        desc = "LSP Definitions / references / ... (Trouble)",
+      },
+      {
+        "<leader>xL",
+        "<cmd>Trouble loclist toggle<cr>",
+        desc = "Location List (Trouble)",
+      },
+      {
+        "<leader>xQ",
+        "<cmd>Trouble qflist toggle<cr>",
+        desc = "Quickfix List (Trouble)",
+      },
+    },
+  },
+
+  {
+    "mikavilpas/yazi.nvim",
+    version = "*",
+    event = "VeryLazy",
+    dependencies = {
+      { "nvim-lua/plenary.nvim", lazy = true },
+    },
+    keys = {
+      {
+        "<leader>yf",
+        mode = { "n", "v" },
+        "<cmd>Yazi<cr>",
+        desc = "Open yazi at the current file",
+      },
+      {
+        "<leader>yw",
+        "<cmd>Yazi cwd<cr>",
+        desc = "Open yazi in current working directory",
+      },
+      {
+        "<leader>yt",
+        "<cmd>Yazi toggle<cr>",
+        desc = "Toggle yazi session",
+      },
+    },
+    ---@type YaziConfig | {}
+    opts = {
+      open_for_directories = true,
+      keymaps = {
+        show_help = "?",
+        open_file_in_vertical_split = "<c-v>",
+        open_file_in_horizontal_split = "<c-s>",
+        open_file_in_tab = "<c-t>",
+        grep_in_directory = "<c-g>",
+        replace_in_directory = "<c-r>",
+        cycle_open_buffers = "<tab>",
+        copy_relative_path_to_selected_files = "<c-y>",
+        send_to_quickfix_list = "<c-q>",
+        change_working_directory = "<c-\\>",
+        open_and_pick_window = "<c-o>",
+      },
+    },
+    init = function()
+      vim.g.loaded_netrwPlugin = 1
+    end,
+  },
+}
